Third pass at code, with containerization.
Changes:

* More tweaks to petl.ini LDAP query configuration.
* Reorganization of project hierarchy to something more modular.
* Moved Python module into a "python" subtree.
* Addition of a "docker" subtree with containerization configuration.
* Addition of log-to-Slack functionality to Python script.
* Complete rewrite of the project README file.

This commit gets us most or all of the way to version 1.2.0 of
the code, and shifts from delivery as a Python module, to delivery
of a container image embedding that Python module.

In general, you should be able to build the Python module, then
build the Docker container image, then instantiate that image into
a Docker server of your choice, and have everything Just Work.

Gory details are in the revised README file.